Cleanup and componentize code
