+++ src/pages/index.tsx
@@ -1,16 +1,9 @@
 import React, { useState, useEffect } from "react";
-import Link from "next/link";
 import { API, graphqlOperation } from "aws-amplify";
-import { listRoutines, listUsers } from "../graphql/queries";
+import { listRoutines } from "../graphql/queries";
 import { GetServerSideProps } from "next";
-import Button from "@material-ui/core/Button";
-import Card from "@material-ui/core/Card";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import CardMedia from "@material-ui/core/CardMedia";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
 import Container from "@material-ui/core/Container";
 import RoutineCard from "../components/RoutineCard";
@@ -24,7 +17,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-function Index({ routinesList, errors }: { routinesList: any; errors: any[] }) {
+function Index({ routinesList }: { routinesList: any; errors: any[] }) {
   const [routines, setRoutines] = useState(routinesList);
   const classes = useStyles();
 
@@ -51,15 +44,12 @@ function Index({ routinesList, errors }: { routinesList: any; errors: any[] }) {
   );
 }
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  // Make a call to the graphQL API here to get all todos from a todolist
-  // The todo list will be 'global'
+export const getServerSideProps: GetServerSideProps = async () => {
   const result = (await API.graphql(graphqlOperation(listRoutines))) as {
     data: ListRoutinesQuery;
     errors: any[];
   };
 
-  // If there's no errors return the todos as an array of
   if (!result.errors) {
     return {
       props: {
@@ -74,7 +64,6 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
       routinesList: [],
       errors: result.errors,
     },
-    // revalidate: 1,
   };
 };
