+++ src/types/CognitoUserAttributes.tsx
@@ -1,0 +1,9 @@
+type CognitoUserAttributes = {
+  email: string;
+  email_verified: boolean;
+  phone_number: string;
+  phone_number_verified: boolean;
+  sub: string;
+};
+
+export default CognitoUserAttributes;
